// ev-backoffice/client/src/app/core/models/government-fee.enum.ts
export enum GovernmentFeeValues {
  biometricServiceFee = 'biometricServiceFee',
  i129f = 'i129f',
  i130 = 'i130',
  i360 = 'i360',
  i485 = 'i485',
  i485_14 = 'i485_14',
  i600_600a = 'i600_600a',
  i601 = 'i601',
  i601a = 'i601a',
  i751 = 'i751',
  i765 = 'i765',
  n400 = 'n400',
  n600_n600k = 'n600_n600k',
}